Use min/max templated functions from std (part 2).

+++ src/DSUtil/MediaTypeEx.cpp
@@ -723,19 +723,19 @@ void CMediaTypeEx::Dump(CAtlList<CString>& sl)
         for (ptrdiff_t i = 0, j = (cbFormat + 15) & ~15; i < j; i += 16) {
             str.Format(_T("%04Ix:"), i);
 
-            for (ptrdiff_t k = i, l = std::min(i + 16, (int)cbFormat); k < l; k++) {
+            for (ptrdiff_t k = i, l = std::min(i + 16, (ptrdiff_t)cbFormat); k < l; k++) {
                 CString byte;
                 byte.Format(_T("%c%02x"), fmtsize > 0 && fmtsize == k ? '|' : ' ', pbFormat[k]);
                 str += byte;
             }
 
-            for (ptrdiff_t k = std::min(i + 16, (int)cbFormat), l = i + 16; k < l; k++) {
+            for (ptrdiff_t k = std::min(i + 16, (ptrdiff_t)cbFormat), l = i + 16; k < l; k++) {
                 str += _T("   ");
             }
 
             str += ' ';
 
-            for (ptrdiff_t k = i, l = std::min(i + 16, (int)cbFormat); k < l; k++) {
+            for (ptrdiff_t k = i, l = std::min(i + 16, (ptrdiff_t)cbFormat); k < l; k++) {
                 unsigned char c = (unsigned char)pbFormat[k];
                 CStringA ch;
                 ch.Format("%c", c >= 0x20 ? c : '.');
